code formatting - moved sonar data sync from codequality to sonar webhook
