+++ src/main/java/com/capitalone/dashboard/webhook/sonarqube/SonarQubeHookController.java
@@ -1,6 +1,7 @@
 package com.capitalone.dashboard.webhook.sonarqube;
 
 import com.capitalone.dashboard.misc.HygieiaException;
+import com.capitalone.dashboard.request.SonarDataSyncRequest;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,9 +11,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.validation.Valid;
 import java.net.MalformedURLException;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static org.springframework.web.bind.annotation.RequestMethod.GET;
 import static org.springframework.web.bind.annotation.RequestMethod.POST;
 
 @RestController
@@ -32,5 +35,10 @@ public class SonarQubeHookController {
         return ResponseEntity
                 .status(HttpStatus.CREATED)
                 .body(response);
+    }
+
+    @RequestMapping(value = "/sonarqube/data-sync", method = GET, produces = APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> codeQualityStaticAnalysisDataSync(@Valid SonarDataSyncRequest request) throws HygieiaException {
+        return sonarQubeHookService.syncData(request);
     }
 }
