+++ src/main/java/com/capitalone/dashboard/webhook/sonarqube/SonarQubeHookServiceImpl.java
@@ -1,11 +1,22 @@
 package com.capitalone.dashboard.webhook.sonarqube;
 
+import com.capitalone.dashboard.client.RestClient;
+import com.capitalone.dashboard.collector.RestOperationsSupplier;
 import com.capitalone.dashboard.misc.HygieiaException;
-import com.capitalone.dashboard.model.*;
 import com.capitalone.dashboard.repository.CodeQualityRepository;
 import com.capitalone.dashboard.repository.CollectorRepository;
+import com.capitalone.dashboard.repository.ComponentRepository;
 import com.capitalone.dashboard.repository.SonarProjectRepository;
-
+import com.capitalone.dashboard.model.SonarProject;
+import com.capitalone.dashboard.model.Collector;
+import com.capitalone.dashboard.model.CollectorType;
+import com.capitalone.dashboard.model.CodeQuality;
+import com.capitalone.dashboard.model.CodeQualityMetric;
+import com.capitalone.dashboard.model.CodeQualityType;
+import com.capitalone.dashboard.model.Component;
+import com.capitalone.dashboard.model.CollectorItem;
+import com.capitalone.dashboard.request.SonarDataSyncRequest;
+import com.google.common.base.Strings;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import org.apache.commons.logging.Log;
@@ -15,15 +26,21 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClientException;
 
-import java.math.BigDecimal;
-import java.net.MalformedURLException;
+import javax.annotation.Nullable;
 import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.List;
+import java.util.Optional;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.concurrent.atomic.AtomicInteger;
 
 @Service
 public class SonarQubeHookServiceImpl implements SonarQubeHookService {
@@ -57,13 +74,17 @@ public class SonarQubeHookServiceImpl implements SonarQubeHookService {
     private final CodeQualityRepository codeQualityRepository;
     private final SonarProjectRepository sonarProjectRepository;
     private final CollectorRepository collectorRepository;
+    private final ComponentRepository componentRepository;
+    private final RestClient restClient = new RestClient(new RestOperationsSupplier());
 
     @Autowired
-    SonarQubeHookServiceImpl( CodeQualityRepository codeQualityRepository, SonarProjectRepository sonarProjectRepository,CollectorRepository collectorRepository)
+    SonarQubeHookServiceImpl( CodeQualityRepository codeQualityRepository, SonarProjectRepository sonarProjectRepository,
+                              CollectorRepository collectorRepository, ComponentRepository componentRepository)
     {
         this.codeQualityRepository = codeQualityRepository;
         this.sonarProjectRepository = sonarProjectRepository;
         this.collectorRepository = collectorRepository;
+        this.componentRepository = componentRepository;
     }
 
     @Override
@@ -268,5 +289,132 @@ public class SonarQubeHookServiceImpl implements SonarQubeHookService {
             }
         }
         return 0;
+    }
+
+    /**
+     * Sync code quality static analysis data from one server to the another
+     * @param request
+     * @return
+     * @throws HygieiaException
+     */
+    public ResponseEntity<String> syncData(SonarDataSyncRequest request) throws HygieiaException {
+
+        String getVersionEpt = "/api/server/version";
+        List<SonarProject> updatedProjects = new ArrayList<>();
+        String from = request.getSyncFrom();
+        String to = request.getSyncTo();
+        boolean isSync = request.getIsSync();
+        final AtomicInteger index = new AtomicInteger();
+        final AtomicInteger compIndex = new AtomicInteger();
+        Collector collector;
+
+        try {
+            if (Strings.isNullOrEmpty(from) ||
+                    Strings.isNullOrEmpty(to) ||
+                    !HttpStatus.OK.equals(restClient.makeRestCallGet(from + getVersionEpt).getStatusCode()) ||
+                    !HttpStatus.OK.equals(restClient.makeRestCallGet(to + getVersionEpt).getStatusCode())) {
+                throw new HygieiaException("Invalid arguments...", HygieiaException.INVALID_CONFIGURATION);
+            }
+            collector = collectorRepository.findByName("Sonar");
+            if (collector == null) {
+                throw new HygieiaException("Collector not found", HygieiaException.COLLECTOR_CREATE_ERROR);
+            }
+        } catch (Exception e) {
+            throw new HygieiaException("Invalid arguments...", HygieiaException.INVALID_CONFIGURATION);
+        }
+        List<SonarProject> projects = getSonarProjects(to);
+        projects.stream().forEach(project -> {
+            String projectName = project.getProjectName();
+            Optional<SonarProject> existingSonarProjectOpt = Optional.ofNullable(sonarProjectRepository.findSonarProjectByProjectName(
+                    collector.getId(), from, projectName));
+            existingSonarProjectOpt.ifPresent(eSonarProject -> {
+                eSonarProject.setProjectId(project.getProjectId());
+                eSonarProject.setInstanceUrl(to);
+                updatedProjects.add(eSonarProject);
+                updateComponent(eSonarProject, isSync, compIndex);
+                LOG.info((index.getAndIncrement() + " : " + projectName + "'s project id & instance url updated"));
+            });
+        });
+
+        String math = updatedProjects.size() + "/" + projects.size();
+        String message = math + " sonar collector items and " + compIndex + " dashboard components can be updated";
+        if (isSync) {
+            sonarProjectRepository.save(updatedProjects);
+            message = math + " sonar collector items and " + compIndex + " dashboard components updated";
+        }
+        LOG.info(message);
+        return ResponseEntity.ok(message);
+    }
+
+    /**
+     * Get code quality projects
+     * @param serverUrl
+     * @return List
+     * @throws HygieiaException
+     */
+    private List<SonarProject> getSonarProjects(String serverUrl) throws HygieiaException {
+
+        String getProjectsEpt = serverUrl + "/api/components/search?qualifiers=TRK&ps=500";
+        JSONParser jsonParser = new JSONParser();
+        JSONArray jsonArray = new JSONArray();
+        List<SonarProject> projects = new ArrayList<>();
+
+        try {
+            ResponseEntity<String> response = restClient.makeRestCallGet(getProjectsEpt);
+            if (!response.getStatusCode().equals(HttpStatus.OK)) {
+                throw new HygieiaException(response.getBody(), HygieiaException.INVALID_CONFIGURATION);
+            }
+            JSONObject responseBody = (JSONObject) jsonParser.parse(response.getBody());
+            Optional<Object> pagingOpt = Optional.ofNullable(responseBody.get("paging"));
+            long totalProjects = pagingOpt.isPresent() ? (Long) ((JSONObject) pagingOpt.get()).get("total") : 0;
+            long pageSize = pagingOpt.isPresent() ? (Long) ((JSONObject) pagingOpt.get()).get("pageSize") : 1;
+            int pages = (int) Math.ceil((double) totalProjects / pageSize);
+
+            if (totalProjects <= pageSize) {
+                jsonArray.addAll((JSONArray) responseBody.get("components"));
+            } else {
+                for (int start = 1; start <= pages; start++) {
+                    String urlFinal = getProjectsEpt + "&p=" + start;
+                    response = restClient.makeRestCallGet(urlFinal);
+                    JSONObject jsonObjectResponse = (JSONObject) jsonParser.parse(response.getBody());
+                    jsonArray.addAll((JSONArray) jsonObjectResponse.get("components"));
+                }
+            }
+        } catch (Exception e) {
+            throw new HygieiaException(e.getMessage(), HygieiaException.INVALID_CONFIGURATION);
+        }
+        jsonArray.forEach(jsonObj -> {
+            JSONObject prjData = (JSONObject) jsonObj;
+            SonarProject project = new SonarProject();
+            project.setInstanceUrl(serverUrl);
+            project.setProjectId((String) Optional.ofNullable(prjData.get("id")).orElse(null));
+            project.setProjectName((String) Optional.ofNullable(prjData.get("name")).orElse(null));
+            projects.add(project);
+        });
+        return projects;
+    }
+
+    /**
+     * Updated the component's code quality collector items' project id and instance url
+     * @param eSonarProject
+     * @param isSync
+     * @param compIndex
+     */
+    private void updateComponent(SonarProject eSonarProject, boolean isSync, @Nullable AtomicInteger compIndex) {
+        List<Component> components = componentRepository.findByCodeQualityCollectorItems(eSonarProject.getId());
+        List<CollectorItem> codeQualityCollectorItems = new ArrayList<>();
+        components.forEach(component -> {
+            component.getCollectorItems(CollectorType.CodeQuality).forEach(collectorItem -> {
+                if (eSonarProject.getProjectName().equals((String) collectorItem.getOptions().get("projectName"))){
+                    collectorItem.getOptions().put("projectId", eSonarProject.getProjectId());
+                    collectorItem.getOptions().put("instanceUrl", eSonarProject.getInstanceUrl());
+                    if (null != compIndex) { compIndex.getAndIncrement(); }
+                }
+                codeQualityCollectorItems.add(collectorItem);
+            });
+            component.setCollectorItems(Collections.singletonMap(CollectorType.CodeQuality, codeQualityCollectorItems));
+            if (isSync) { componentRepository.save(component); }
+        });
+
     }
 }
