+++ src/test/java/com/capitalone/dashboard/webhook/sonarqube/SonarQubeWebhookTestApi.java
@@ -4,6 +4,7 @@ import com.capitalone.dashboard.client.RestClient;
 import com.capitalone.dashboard.model.Collector;
 import com.capitalone.dashboard.repository.CodeQualityRepository;
 import com.capitalone.dashboard.repository.CollectorRepository;
+import com.capitalone.dashboard.repository.ComponentRepository;
 import com.capitalone.dashboard.repository.SonarProjectRepository;
 import com.capitalone.dashboard.util.Supplier;
 import com.google.common.io.Resources;
@@ -28,6 +29,7 @@ public class SonarQubeWebhookTestApi {
     @Mock private CodeQualityRepository codeQualityRepository;
     @Mock private SonarProjectRepository sonarProjectRepository;
     @Mock private CollectorRepository collectorRepository;
+    @Mock private ComponentRepository componentRepository;
 
     @Mock
     private Supplier<RestOperations> restOperationsSupplier;
@@ -40,7 +42,7 @@ public class SonarQubeWebhookTestApi {
     public void init() {
         RestClient restClientTemp = new RestClient(restOperationsSupplier);
         restClient = Mockito.spy(restClientTemp);
-        sonarQubeHookService = new SonarQubeHookServiceImpl(codeQualityRepository,sonarProjectRepository,collectorRepository);
+        sonarQubeHookService = new SonarQubeHookServiceImpl(codeQualityRepository,sonarProjectRepository,collectorRepository, componentRepository);
     }
 
     @Test
